Move Predicate to IR module
