+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/ast/convert/plannerQuery/groupInequalityPredicates.scala
@@ -19,10 +19,10 @@
  */
 package org.neo4j.cypher.internal.compiler.v3_2.ast.convert.plannerQuery
 
-import org.neo4j.cypher.internal.frontend.v3_2.ast.{AndedPropertyInequalities, Variable, InequalityExpression, Property}
+import org.neo4j.cypher.internal.frontend.v3_2.ast.{AndedPropertyInequalities, InequalityExpression, Property, Variable}
 import org.neo4j.cypher.internal.frontend.v3_2.helpers.NonEmptyList
-import org.neo4j.cypher.internal.compiler.v3_2.planner.Predicate
 import org.neo4j.cypher.internal.frontend.v3_2.helpers.NonEmptyList._
+import org.neo4j.cypher.internal.ir.v3_2.Predicate
 
 // This transforms
 //
