+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/OptionalMatchRemover.scala
@@ -27,7 +27,7 @@ import org.neo4j.cypher.internal.frontend.v3_2.phases.CompilationPhaseTracer.Com
 import org.neo4j.cypher.internal.frontend.v3_2.phases.CompilationPhaseTracer.CompilationPhase.LOGICAL_PLANNING
 import org.neo4j.cypher.internal.frontend.v3_2.phases.Phase
 import org.neo4j.cypher.internal.frontend.v3_2.{InputPosition, Rewriter, topDown}
-import org.neo4j.cypher.internal.ir.v3_2.{AggregatingQueryProjection, IdName, PatternRelationship}
+import org.neo4j.cypher.internal.ir.v3_2.{AggregatingQueryProjection, IdName, PatternRelationship, Predicate}
 
 import scala.annotation.tailrec
 import scala.collection.mutable.ListBuffer
