+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/steps/planShortestPaths.scala
@@ -22,14 +22,14 @@ package org.neo4j.cypher.internal.compiler.v3_2.planner.logical.steps
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.idp.expandSolverStep
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans.{DoNotIncludeTies, IncludeTies, LogicalPlan}
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.{Ascending, LogicalPlanningContext}
-import org.neo4j.cypher.internal.compiler.v3_2.planner.{Predicate, QueryGraph}
+import org.neo4j.cypher.internal.compiler.v3_2.planner.QueryGraph
 import org.neo4j.cypher.internal.frontend.v3_2.ast._
 import org.neo4j.cypher.internal.frontend.v3_2.ast.functions.{Length, Nodes}
 import org.neo4j.cypher.internal.frontend.v3_2.ast.rewriters.projectNamedPaths
 import org.neo4j.cypher.internal.frontend.v3_2.helpers.FreshIdNameGenerator
 import org.neo4j.cypher.internal.frontend.v3_2.notification.ExhaustiveShortestPathForbiddenNotification
 import org.neo4j.cypher.internal.frontend.v3_2.{ExhaustiveShortestPathForbiddenException, InternalException}
-import org.neo4j.cypher.internal.ir.v3_2.{IdName, ShortestPathPattern}
+import org.neo4j.cypher.internal.ir.v3_2.{IdName, Predicate, ShortestPathPattern}
 
 case object planShortestPaths {
