+++ community/cypher/cypher-compiler-3.2/src/test/scala/org/neo4j/cypher/internal/compiler/v3_2/ast/convert/plannerQuery/PatternExpressionConverterTest.scala
@@ -20,11 +20,11 @@
 package org.neo4j.cypher.internal.compiler.v3_2.ast.convert.plannerQuery
 
 import org.neo4j.cypher.internal.compiler.v3_2.ast.convert.plannerQuery.ExpressionConverters._
-import org.neo4j.cypher.internal.compiler.v3_2.planner.{LogicalPlanningTestSupport, Predicate, Selections}
+import org.neo4j.cypher.internal.compiler.v3_2.planner.{LogicalPlanningTestSupport, Selections}
 import org.neo4j.cypher.internal.frontend.v3_2.ast.{Expression, Variable}
 import org.neo4j.cypher.internal.frontend.v3_2.test_helpers.CypherFunSuite
 import org.neo4j.cypher.internal.frontend.v3_2.{SemanticDirection, ast}
-import org.neo4j.cypher.internal.ir.v3_2.{IdName, PatternRelationship, SimplePatternLength}
+import org.neo4j.cypher.internal.ir.v3_2.{IdName, PatternRelationship, Predicate, SimplePatternLength}
 
 class PatternExpressionConverterTest extends CypherFunSuite with LogicalPlanningTestSupport {
