+++ community/cypher/cypher-compiler-3.2/src/test/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/cardinality/ExpressionSelectivityCalculatorTest.scala
@@ -20,14 +20,14 @@
 package org.neo4j.cypher.internal.compiler.v3_2.planner.logical.cardinality
 
 import org.mockito.Mockito.when
-import org.neo4j.cypher.internal.compiler.v3_2.planner.{Predicate, Selections}
+import org.neo4j.cypher.internal.compiler.v3_2.planner.Selections
 import org.neo4j.cypher.internal.compiler.v3_2.spi.GraphStatistics
 import org.neo4j.cypher.internal.frontend.v3_2.ast._
 import org.neo4j.cypher.internal.frontend.v3_2.helpers.NonEmptyList
 import org.neo4j.cypher.internal.frontend.v3_2.symbols._
 import org.neo4j.cypher.internal.frontend.v3_2.test_helpers.CypherFunSuite
 import org.neo4j.cypher.internal.frontend.v3_2.{InputPosition, LabelId, PropertyKeyId, SemanticTable}
-import org.neo4j.cypher.internal.ir.v3_2.{Cardinality, IdName, Selectivity}
+import org.neo4j.cypher.internal.ir.v3_2.{Cardinality, IdName, Predicate, Selectivity}
 
 class ExpressionSelectivityCalculatorTest extends CypherFunSuite with AstConstructionTestSupport {
